Extract swipe-picker.js module
